backend/requirements.txt
fastapi
uvicorn
python-multipart
sse-starlette
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
loguru
pipecat-ai[websocket,elevenlabs,silero,openai,deepgram,runner]
twilio
torch